+++ JamGame/Assets/Scripts/AnimatorsSwitcher.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd9e9183aea348c488697fe85c089277
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
