+++ JamGame/Assets/Scripts/AnimatorsSwitcher/Model.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 564bbdba624e2814b9744e85fedd8441
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
