+++ JamGame/Assets/Scripts/Level/TransitionPanel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b968c641653597c489d3130aebd0a0d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
