+++ JamGame/Assets/Scripts/Level/TransitionPanel/Model.cs
@@ -1,0 +1,22 @@
+﻿using UnityEngine;
+using UnityEngine.Events;
+
+namespace Level.TransitionPanel
+{
+    [AddComponentMenu("Level.TransitionPanel.Model")]
+    public class Model : MonoBehaviour
+    {
+        [SerializeField, HideInInspector]
+        private string panelText;
+        public UnityEvent<string> TextChange;
+        public string PanelText
+        {
+            get => panelText;
+            set
+            {
+                panelText = value;
+                TextChange?.Invoke(panelText);
+            }
+        }
+    }
+}
