+++ JamGame/Assets/Scripts/Level/TransitionPanel/Model.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 370bdfc5155fa8b44831f7f78e34d463
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
