+++ JamGame/Assets/Scripts/Level/TransitionPanel/View.cs
@@ -1,0 +1,17 @@
+﻿using TMPro;
+using UnityEngine;
+
+namespace Level.TransitionPanel
+{
+    [AddComponentMenu("Level.TransitionPanel.View")]
+    public class View : MonoBehaviour
+    {
+        [SerializeField]
+        private TMP_Text textOnPanel;
+
+        public void UpdateText(string text)
+        {
+            textOnPanel.text = text;
+        }
+    }
+}
