+++ JamGame/Assets/Scripts/Level/TransitionPanel/View.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26bc5e9f4c5a18c4394f5bc2cf9b9aed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
