--- JamGame/Assets/Scripts/UI.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:JamGame/Assets/Scripts/UI.meta
-guid: 39142c27be09ed140bb00ec9c2b6f74b
-folderAsset: yes
-========
-guid: 2d852de2dd9de6b47a3e076d3bcf6063
->>>>>>>> 6ea48eff09e3a1650478f64384eb2ce7281f78c6:JamGame/Assets/Scenes/PathFinding.unity.meta
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
