--- JamGame/Assets/Scripts/UI/Overlays.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1218d4d105b28074485629a08fb9e7ea
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
