--- JamGame/Assets/Scripts/UI/Overlays/OverlaySelector.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8e2b9d3632791214d96177d4d70b3101
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
