--- JamGame/Assets/Scripts/UI/TransitionPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 26bc5e9f4c5a18c4394f5bc2cf9b9aed
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
