--- JamGame/Assets/Scripts/UI/UIController.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0a627cb8612150248b796d51b070f2e3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
