--- JamGame/Assets/Scripts/UI/UIController/AnimatorListDrawerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea822bf4840fa0042b871857816b91a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
