--- JamGame/Assets/Scripts/UI/UIController/AnimatorsSwitcher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40c5b5298e457d74a97c3a5224845dc7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
